Sanitized the current translation.
Many changes that as a consequence had various warnings being printed
out when compiling the locale.  Sometimes the spacing between elements
caused the access keys to be recognized in error.  Sometimes there
were stale .po files which are no longer present in the templates.

The "encoding errors" in the *.it.po files remain, and they are from
unknown source.

The .it files do not seem to be used in FF3 so this should not pose a
problem for the l10n of 3.


git-svn-id: svn+ssh://cowfarm/usr/local/src/svnroot/mozilla@536 a9a49e30-ccf7-0310-8831-bfded88c6097